requirements.txt
netmiko
paramiko
asyncio
requests
gns3fy
pyyaml
colorama
flask
flask-cors
